+++ src/vm.cpp
@@ -1683,10 +1683,6 @@ bool VM::handleException(DSState &state, bool forceUnwind) {
     return false;
 }
 
-#ifdef CODE_COVERAGE
-extern "C" void __gcov_flush(); // for coverage reporting
-#endif
-
 DSValue VM::startEval(DSState &state, EvalOP op, DSError *dsError) {
     DSValue value;
     const unsigned int oldLevel = state.subshellLevel;
@@ -1716,14 +1712,7 @@ DSValue VM::startEval(DSState &state, EvalOP op, DSError *dsError) {
     }
 
     if(subshell) {
-#ifdef CODE_COVERAGE
-        /*
-         * after call _exit(), not write coverage information due to skip atexit handler.
-         * in order to write coverage information, manually call __gcove_flush()
-         */
-        __gcov_flush();
-#endif
-        _exit(state.getMaskedExitStatus());
+        terminate(state.getMaskedExitStatus());
     }
 
     if(hasFlag(op, EvalOP::COMMIT)) {